+++ include/function.h
@@ -1,7 +1,7 @@
 /**
  * @file function.h
  *
- * @brief `Nst_FuncObj` interface
+ * @brief `Func` object interface
  *
  * @author TheSilvered
  */
@@ -12,47 +12,11 @@
 #include "compiler.h"
 #include "map.h"
 
-/* Casts `ptr` to `Nst_FuncObj *`. */
-#define FUNC(ptr) ((Nst_FuncObj *)(ptr))
-/**
- * @brief Alias for `_Nst_func_set_vt` that casts `func` to `Nst_FuncObj *` and
- * `map` to `Nst_MapObj *`.
- */
-#define Nst_func_set_vt(func, map)                                            \
-    _Nst_func_set_vt((Nst_FuncObj *)(func), (Nst_MapObj *)(map))
+#define Nst_FUNC_IS_C(func) ((func)->flags & Nst_FLAG_FUNC_IS_C)
 
 #ifdef __cplusplus
 extern "C" {
 #endif // !__cplusplus
-
-
-/**
- * The union representing the body of a function object.
- *
- * @param bytecode: the body is an instruction list
- * @param c_func: the body is a C function
- */
-NstEXP typedef union _Nst_FuncBody {
-    Nst_InstList *bytecode;
-    Nst_Obj *(*c_func)(usize arg_num, Nst_Obj **args);
-} Nst_FuncBody;
-
-/**
- * The structure representing a Nest function object.
- *
- * @param body: the body of the function
- * @param args: the array of names of the arguments
- * @param arg_num: the maximum number of arguments
- * @param mod_globals: the global variable table when the function was defined
- */
-NstEXP typedef struct _Nst_FuncObj {
-    Nst_OBJ_HEAD;
-    Nst_GGC_HEAD;
-    Nst_FuncBody body;
-    Nst_Obj **args;
-    isize arg_num;
-    Nst_MapObj *mod_globals;
-} Nst_FuncObj;
 
 /**
  * Creates a new function object with an instruction-list body.
@@ -86,14 +50,28 @@ NstEXP Nst_Obj *NstC Nst_func_new_c(usize arg_num, Nst_NestCallable cbody);
  * @param func: the function to change the `mod_globals` field of
  * @param map: the map to set as the new value
  */
-NstEXP void NstC _Nst_func_set_vt(Nst_FuncObj *func, Nst_MapObj *map);
+NstEXP void NstC Nst_func_set_vt(Nst_Obj *func, Nst_MapObj *map);
 
-/* Traverse function for `Nst_FuncObj`. */
-NstEXP void NstC _Nst_func_traverse(Nst_FuncObj *func);
-/* Destructor for `Nst_FuncObj`. */
-NstEXP void NstC _Nst_func_destroy(Nst_FuncObj *func);
+/* Gets the number of arguments a function takes. */
+NstEXP usize NstC Nst_func_arg_num(Nst_Obj *func);
+/* Gets the argument names as a list of objects. */
+NstEXP Nst_Obj **NstC Nst_func_args(Nst_Obj *func);
+/* Gets the body of a C-function wrapper. */
+NstEXP Nst_NestCallable NstC Nst_func_c_body(Nst_Obj *func);
+/* Gets the body of a Nest function. */
+NstEXP Nst_InstList *NstC Nst_func_nest_body(Nst_Obj *func);
+/**
+ * @brief Gets the `_globals_` variable map of a function. No reference is
+ * added. It may be `NULL`.
+ */
+NstEXP Nst_MapObj *NstC Nst_func_mod_globals(Nst_Obj *func);
 
-/* The flags for `Nst_FuncObj`. */
+/* Traverse function for `Func` objects. */
+NstEXP void NstC _Nst_func_traverse(Nst_Obj *func);
+/* Destructor for `Func` objects. */
+NstEXP void NstC _Nst_func_destroy(Nst_Obj *func);
+
+/* Flags for `Func` objects. */
 NstEXP typedef enum _Nst_FuncFlags {
     Nst_FLAG_FUNC_IS_C = Nst_FLAG(1)
 } Nst_FuncFlags;